Export CheckPlot and move to internal
