+++ vg/vg_test.go
@@ -6,7 +6,6 @@ package vg_test
 
 import (
 	"bytes"
-	"flag"
 	"fmt"
 	"io/ioutil"
 	"log"
@@ -18,8 +17,6 @@ import (
 	"github.com/gonum/plot/plotter"
 	"github.com/gonum/plot/vg"
 )
-
-var generateTestData = flag.Bool("regen", false, "Uses the current state to regenerate the test data.")
 
 // TestLineWidth tests output against test images generated by
 // running tests with -tag good.
@@ -58,7 +55,7 @@ func TestLineWidth(t *testing.T) {
 			name := filepath.Join(".", "testdata", fmt.Sprintf("width_%v.%s", w, typ))
 
 			// Recreate Golden images.
-			if *generateTestData {
+			if *cmpimg.GenerateTestData {
 				err = p.Save(width, height, name)
 				if err != nil {
 					log.Fatalf("failed to save %q: %v", name, err)
